requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.28.1
google-generativeai==0.3.0
requests==2.31.0
